重构词典制作代码

+++ src/main/java/zx/soft/naive/bayes/mapred/db/DbToWordsReducer.java
@@ -2,24 +2,24 @@ package zx.soft.naive.bayes.mapred.db;
 
 import java.io.IOException;
 
-import org.apache.hadoop.io.LongWritable;
+import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Reducer;
 
-public class DbToWordsReducer extends Reducer<Text, LongWritable, Text, LongWritable> {
+public class DbToWordsReducer extends Reducer<Text, IntWritable, Text, IntWritable> {
 
-	private static long sum = 0;
+	private static int sum = 0;
 
 	@Override
-	protected void reduce(Text word, Iterable<LongWritable> values, Context context) throws IOException,
+	protected void reduce(Text word, Iterable<IntWritable> values, Context context) throws IOException,
 			InterruptedException {
 
 		sum = 0;
 
-		for (LongWritable value : values) {
+		for (IntWritable value : values) {
 			sum += value.get();
 		}
-		context.write(word, new LongWritable(sum));
+		context.write(word, new IntWritable(sum));
 	}
 
 }
