+++ src/main/java/zx/soft/naive/bayes/mapred/db/MergeWordsMapper.java
@@ -2,19 +2,14 @@ package zx.soft.naive.bayes.mapred.db;
 
 import java.io.IOException;
 
-import org.apache.hadoop.io.LongWritable;
+import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Mapper;
-import org.apache.hadoop.mapreduce.lib.input.FileSplit;
 
-import zx.soft.naive.bayes.analyzer.AnalyzerTool;
-
-
-public class MergeWordsMapper extends Mapper<Text, LongWritable,  Text, LongWritable> {
+public class MergeWordsMapper extends Mapper<Text, IntWritable, Text, IntWritable> {
 
 	@Override
-	public void map(Text key, LongWritable value, Context context) throws InterruptedException, IOException {
-
+	public void map(Text key, IntWritable value, Context context) throws InterruptedException, IOException {
 		context.write(key, value);
 	}
